GCW-3544 Create types for api req, response
